+++ R/virtualenv.R
@@ -35,19 +35,20 @@
 #'   environments?
 #'
 #' @param python The path to a Python interpreter, to be used with the created
-#'   virtual environment. When `NULL`, the Python interpreter associated with
-#'   the current session will be used.
+#'   virtual environment. This can also accept a version constraint like
+#'   `"3.10"`, which is passed on to `virtualenv_starter()` to find a suitable
+#'   python binary.
 #'
 #' @param force Boolean; force recreating the environment specified by
-#'   `envname`, even if it already exists. If `TRUE`, the previous enviroment is
-#'   first deleted and recreated. Otherwise, if `FALSE`, the path to the
-#'   existing environment is returned.
+#'   `envname`, even if it already exists. If `TRUE`, the pre-existing
+#'   environment is first deleted and then recreated. Otherwise, if `FALSE` (the
+#'   default), the path to the existing environment is returned.
 #'
-#' @param version,python_version The version of Python to use when creating a
-#'   virtual environment. Python installations will be searched for using
-#'   [`virtualenv_starter()`]. This can a specific version, like `"3.9"` or
-#'   `"3.9.3`, or a comma separated list of version constraints, like `">=3.8"`,
-#'   or `"<=3.11,!=3.9.3,>3.6"`
+#' @param version,python_version (string) The version of Python to use when
+#'   creating a virtual environment. Python installations will be searched for
+#'   using [`virtualenv_starter()`]. This can a specific version, like `"3.9"`
+#'   or `"3.9.3"`, or a comma separated list of version constraints, like
+#'   `">=3.8"`, or `"<=3.11,!=3.9.3,>3.6"`
 #'
 #' @param all If `TRUE`, `virtualenv_starter()` returns a 2-column data frame,
 #'   with column names `path` and `version`. If `FALSE`, only a single path to a
@@ -104,6 +105,8 @@ virtualenv_create <- function(
   setuptools_version   = getOption("reticulate.virtualenv.setuptools_version", default = NULL),
   extra                = getOption("reticulate.virtualenv.extra", default = NULL))
 {
+  check_forbidden_install("Python Virtual Environments")
+
   path <- virtualenv_path(envname)
   name <- if (is.null(envname)) path else envname
 
@@ -117,10 +120,18 @@ virtualenv_create <- function(
     }
   }
 
-  if (is.null(python))
-    stop_no_virtualenv_starter(version)
+  # for convenience, also accept a version constraint in the 2nd positional arg
+  # e.g.: virtualenv_create("r-foo", "3.10")
+  if(is.null(version) && is_string(python) &&
+     !grepl("[/\\]", python) && # not a file path
+     grepl("^[0-9.><=!,]+$", python)) # maybe a version constraint
+    python <- virtualenv_starter(python)
 
-  check_can_be_virtualenv_starter(python)
+  else if (is.null(python))
+    python <- virtualenv_starter(version)
+
+
+  check_can_be_virtualenv_starter(python, version)
 
   module <- module %||% virtualenv_module(python)
 
@@ -169,13 +180,18 @@ virtualenv_create <- function(
 
   writef("Done!")
 
+  if (missing(packages) && !is.null(requirements))
+    packages <- NULL
+
   # upgrade pip and friends after creating the environment
   # (since the version bundled with virtualenv / venv may be stale)
   if (!identical(packages, FALSE)) {
     python <- virtualenv_python(envname)
     # first upgrade pip and friends
     writef("Installing packages: pip, wheel, setuptools")
-    pip_install(python, c("pip", "wheel", "setuptools"))
+    pip_install(python, c("pip", "wheel", "setuptools"),
+                # on centos7, system pip too old, no support for --no-user
+                no_user = FALSE)
     packages <- setdiff(packages, c("pip", "wheel", "setuptools"))
     # install requested packages
     if (length(packages)) {
@@ -408,8 +424,16 @@ virtualenv_module <- function(python) {
 
   # if we have one of these modules available, return it
   for (module in modules)
-    if (python_has_module(python, module))
+    if (python_has_module(python, module)) {
+      if(module == "venv" && is_ubuntu() && startsWith(python, "/usr/bin/python")) {
+        # `apt install python3` makes an importable venv module, but not one
+        # capable of actually creating a venv unless python3-venv is installed.
+        # if python3-venv is not installed, move on and maybe discover virtualenv.
+        if (!any(grepl("^python[0-9.]*-venv$", system("dpkg -l", intern = TRUE))))
+          next
+      }
       return(module)
+    }
 
   # virtualenv not available: instruct the user to install
   commands <- stack(mode = "character")
@@ -652,13 +676,13 @@ as_version_constraint_checkers <- function(version) {
 }
 
 
-check_can_be_virtualenv_starter <- function(python) {
+check_can_be_virtualenv_starter <- function(python, version) {
   if(!can_be_virtualenv_starter(python))
-    stop_no_virtualenv_starter()
+    stop_no_virtualenv_starter(version = version, python = python)
 }
 
 can_be_virtualenv_starter <- function(python) {
-  if (!file.exists(python))
+  if (is.null(python) || !file.exists(python))
     return(FALSE)
 
   # get version
@@ -688,23 +712,42 @@ can_be_virtualenv_starter <- function(python) {
 }
 
 
-stop_no_virtualenv_starter <- function(version = NULL) {
+stop_no_virtualenv_starter <- function(version = NULL, python = NULL) {
 
   .msg <- character()
   w <- function(...) .msg <<- c(.msg, paste0(...))
 
   w("Suitable Python installation for creating a venv not found.")
+  if (!is.null(python))
+    w("  Requested Python: ", python)
   if (!is.null(version))
-    w("Requested version constraint: ", version)
+    w("  Requested version constraint: ", version)
+
   w("Please install Python with one of following methods:")
 
   if (is_linux())
-      w("- https://github.com/rstudio/python-builds/")
+    w("- https://github.com/rstudio/python-builds/")
 
   if (!is_linux())
     w("- https://www.python.org/downloads/")
 
   w("- reticulate::install_python(version = '<version>')")
+
+
+  # On linux, if there is an incomplete system python install,
+  # give a hint on how to make it usable
+  python <- python %||% "/usr/bin/python3"
+  if (is_linux() &&
+      startsWith(python, "/usr/bin/python3") &&
+      file.exists(python) &&
+      !python_has_modules(python, c("pip", "venv"))) {
+    if (is_ubuntu())
+      w("- sudo apt install python3-venv")
+    else if (is_fedora())
+      w("- sudo dnf install python3-pip")
+    else
+      w("- Install python3-venv and python3-pip using the system package manager")
+  }
 
   stop(paste0(.msg, collapse = "\n"))